Update test cases.
